执行增删改查

+++ src/main/java/io/mqttpush/mysql/packet/response/ErrorPacket.java
@@ -21,7 +21,7 @@ public class ErrorPacket extends MysqlPacket{
      *
      *
      */
-	byte header;
+	short header;
 
 	int errorCode;
